# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RetailIQ API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/nlp/query": {
      "post": {
        "summary": "nlp_query_v2",
        "operationId": "nlp_query_v2_post",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Response 400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/ai/nlp/query": {
      "post": {
        "summary": "nlp_query_v2",
        "operationId": "nlp_query_v2_post",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Response 400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/ai/recommend": {
      "post": {
        "summary": "recommend_v2",
        "operationId": "recommend_v2_post",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}